fix: stop relying on `isCall`, properly set input/output regs
Fixes crashing `constructCallSiteEntryDIEs`, which attempted to read the
call target of TC shadow stack pseudoinstructions with `isCall`, but
failed to do so because they didn't have any operands.

This commit replaces the uses of `isCall` with `hasSideEffects`. To
prevent incorrect instruction reordering, the pseudoinstructions now
have input and output registers.